added core data no delete

+++ GaugeView/Models/Recordings.swift
@@ -10,13 +10,13 @@ import Foundation
 
 class Recordings {
     
-    var recordings: Int  //filename
+    var recordings: Double  //filename
     var decibels: Float  // filename dB
     //weak var score: Score? // pass/ fail
     var sst: String
     var score: String
     
-    init(recordings: Int, decibels: Float, sst: String, score: String){
+    init(recordings: Double, decibels: Float, sst: String, score: String){
     self.recordings = recordings
     self.decibels = decibels
     self.sst = sst
